requirements.txt
pyshp
psycopg2-binary
numpy>=1.16.5
pandas
openpyxl
geopandas
shapely
scipy
matplotlib
camelot-py[cv]
odfpy
db_fillers
h3
